feat: Add support for executing custom user data
This commit introduces the ability to run a custom script prior to
installing and starting the runner service. Users can now pass custom
script (runs as sudo) to configure their runners with specific requirements before
the runner is fully initialized.

A new input parameter `custom_script` has been added. It accepts a
string value representing the custom script. If this parameter is
provided, the script will be executed before the runner service starts,
allowing for custom configurations or setup tasks to be performed seamlessly.

Signed-off-by: Vibhu Prashar <vibhu.sharma2929@gmail.com>